Extract Vmdb::Loggers::Instrument to its own file
